requirements.txt
requests==2.32.3
markdownify==0.13.1
beautifulsoup4==4.12.3
python-dotenv==1.0.1
tqdm==4.66.5
psycopg[binary]>=3.2.10,<3.3
